+++ ee_plugin/ui/widgets.py
@@ -1,0 +1,89 @@
+from dataclasses import field
+from typing import List, Optional, Tuple, Union
+
+from qgis import gui
+from qgis.PyQt import QtCore
+from qgis.PyQt.QtWidgets import (
+    QDialog,
+    QDialogButtonBox,
+    QFormLayout,
+    QGroupBox,
+    QLayout,
+    QVBoxLayout,
+    QWidget,
+)
+from qgis.gui import QgsCollapsibleGroupBox
+
+
+def DefaultNullQgsDateEdit(
+    *, date: Optional[QtCore.QDate] = None, displayFormat="yyyy-MM-dd", **kwargs
+) -> gui.QgsDateEdit:
+    """Build a QgsDateEdit widget, with null default capability."""
+    # NOTE: Specifying a displayFormat guarantees that the date will be formatted as
+    # expected across different runtime environments.
+    d = gui.QgsDateEdit(**kwargs, displayFormat=displayFormat)
+    # NOTE: It would be great to remove this helper and just use the built-in QgsDateEdit
+    # class but at this time it's not clear how to make a DateEdit widget that initializes
+    # with a null value. This is a workaround.
+    if date is None:
+        d.clear()
+    else:
+        d.setDate(date)
+    return d
+
+
+def build_form_group_box(
+    *,
+    rows: List[
+        Union[Tuple[Union[QWidget, str], Union[QWidget, QLayout]], QWidget, QLayout]
+    ] = field(default_factory=list),
+    collapsable: bool = False,
+    **kwargs,
+) -> Union[QGroupBox, QgsCollapsibleGroupBox]:
+    """
+    A group box with a form layout.
+    """
+    gb = QGroupBox(**kwargs) if not collapsable else QgsCollapsibleGroupBox(**kwargs)
+    layout = QFormLayout()
+    gb.setLayout(layout)
+
+    for row in rows:
+        if isinstance(row, (QWidget, QLayout)):
+            row = [row]
+        layout.addRow(*row)
+
+    return gb
+
+
+def build_vbox_dialog(
+    widgets: List[QWidget] = field(default_factory=list),
+    show: bool = True,
+    **kwargs,
+) -> QDialog:
+    """
+    Build a dialog with a vertical layout and configured standard buttons.
+    """
+    dialog = QDialog(**kwargs)
+
+    # Configure dialog layout
+    main_layout = QVBoxLayout(dialog)
+    dialog.setLayout(main_layout)
+
+    # Add widgets to the dialog
+    for widget in widgets:
+        main_layout.addWidget(widget)
+
+    # Add OK/Cancel buttons
+    main_layout.addWidget(
+        QDialogButtonBox(
+            standardButtons=QDialogButtonBox.Cancel | QDialogButtonBox.Ok,
+            accepted=dialog.accept,
+            rejected=dialog.reject,
+        )
+    )
+
+    # Show the dialog on screen
+    if show:
+        dialog.show()
+
+    return dialog
